requirements.txt
Flask
numpy
alpaca-trade-api
python-dotenv